Refactored base class out of HeatStructure2DCouplerBC
Refs #21688
